:wrench: refactor (Evolution): Move into its own file.
